% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{downloadRR53Data}
\alias{downloadRR53Data}
\title{Download the data from PNAS and save it locally as a file}
\usage{
downloadRR53Data(outfile)
}
\arguments{
\item{outfile}{name of the file to be stored locally}
}
\value{
No return value.
}
\description{
Download the data from PNAS and save it locally as a file
}

