% Generated by roxygen2 (4.0.2): do not edit by hand
\name{generateVCMatrix}
\alias{generateVCMatrix}
\title{Generate variance-covariance matrix for gene expression data}
\usage{
generateVCMatrix(dat)
}
\arguments{
\item{dat}{matrix of raw gene expression data}
}
\value{
Estimated variance-covariance matrix
}
\description{
Generate variance-covariance matrix for gene expression data
}

