% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{getGeneResids}
\alias{getGeneResids}
\title{Compute the residual gene expression values}
\usage{
getGeneResids(x)
}
\arguments{
\item{x}{Data frame containing gene values for a single gene along with predictor values}
}
\value{
Data frame with new residualized gene values.
}
\description{
Calculates residual gene expression values for a single gene after
controlling for the 15 control factors in Fredrickson et al.
}

