% Generated by roxygen2 (4.0.2): do not edit by hand
\name{meanAbsCor}
\alias{meanAbsCor}
\title{Compute the mean absolute correlation for a correlation matrix.}
\usage{
meanAbsCor(x)
}
\arguments{
\item{x}{Matrix of data}
}
\value{
Mean absolute correlation for the matrix.
}
\description{
This function computes the correlation matrix using \code{\link{cor}},
and then calculates the mean absolute pairwise correlation in the
upper triangle of the matrix.
}

