% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{readData}
\alias{readData}
\title{Read in the data from Fredrickson et al.}
\usage{

  readData(file = "http://www.pnas.org/content/suppl/2014/08/21/1407057111.DCSupplemental/pnas.1407057111.sd02.csv")
}
\arguments{
\item{file}{Name of the local CSV file.  If left unspecified, defaults to download the data from PNAS.}
}
\value{
A dataframe, with fields as described by Brown et al.
}
\description{
This function reads in the data from the PNAS website
or from a downloaded CSV file and tags each row with
a subject ID number (SID).
}

