% Generated by roxygen2 (4.0.2): do not edit by hand
\name{readTidyData}
\alias{readTidyData}
\title{Read in and tidy the data from Fredrickson et al.}
\usage{

  readTidyData(file = "http://www.pnas.org/content/suppl/2014/08/21/1407057111.DCSupplemental/pnas.1407057111.sd02.csv")
}
\arguments{
\item{file}{Name of the local CSV file.  If left unspecified, defaults to download the data from PNAS.}
}
\value{
a list with four elements: Predictors, Genetics, Psychometrics, FactorAssignment, GeneSigns.
}
\description{
This downloads the data from the PNAS website
(or from a local CSV file) and performs some basic
preprocessing to create 'tidy' data.frames for
easy analysis.
}

