% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rr53}
\alias{rr53}
\title{Run 53 repeated regressions as in Fredrickson et al.}
\usage{
rr53(gedata, preds, geneSigns, calibrate = TRUE, joinby = "SID",
  resids = FALSE)
}
\arguments{
\item{gedata}{Genetic data}

\item{preds}{Psychometric data (predictors)}

\item{geneSigns}{Dataframe with contrast codes for genes}

\item{calibrate}{Use eudaimonic/hedonic z-scores from Fredrickson et al. (default \code{TRUE}). Set to FALSE to use Brown et al.'s predictors.}

\item{joinby}{On which field to join the Psychometric / Genetic data (this should be left as is)}

\item{resids}{Whether to run the RR53 on the residualized gene data (\code{TRUE}) or on the raw data (\code{FALSE}=default).}
}
\value{
A dataframe containing the estimated beta weights for the eudaimonic factor.
}
\description{
Run 53 repeated regressions as in Fredrickson et al.
}

