% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rr53Resids}
\alias{rr53Resids}
\title{Run the RR53 analysis on the gene expression residuals}
\usage{
rr53Resids(gdat, pdat, gSigns)
}
\arguments{
\item{gdat}{Data frame with residualized gene expression data}

\item{pdat}{Data frame with predictors}

\item{gSigns}{Data frame with gene contrast codes}
}
\value{
t-value for the one-sample test (bootstrapped SEs)
}
\description{
Run the RR53 analysis on the gene expression residuals
}

