% Generated by roxygen2 (4.0.2): do not edit by hand
\name{runLM}
\alias{runLM}
\title{Fit a linear model to expression data for a single gene}
\usage{
runLM(gdat, pdat, calibrate = TRUE, joinby = "SID", resids = FALSE)
}
\arguments{
\item{gdat}{Data frame containing expression data for the gene}

\item{pdat}{Data frame contining predictors}

\item{calibrate}{Use Fredrickson et al.'s predictors (\code{TRUE}) or Brown et al.'s (\code{FALSE})}

\item{joinby}{What field to join genetic data with psychometric data (leave as is)}

\item{resids}{Whether to use residualized genetic data or raw data}
}
\value{
A data frame with field \code{Pred} naming the predictor, and \code{x} containing its value.
}
\description{
Fit a linear model to expression data for a single gene
}

