% Generated by roxygen2 (4.0.2): do not edit by hand
\name{shuffleAndRecomputeMAC}
\alias{shuffleAndRecomputeMAC}
\title{Compute MAC after randomly shuffling columns}
\usage{
shuffleAndRecomputeMAC(x)
}
\arguments{
\item{x}{Matrix of gene expression data (rows=subjects, columns=genes)}
}
\value{
Mean absolute correlation for the shuffled matrix
}
\description{
Randomly shuffle the columns of the matrix and then recompute the
mean absolute correlation using \code{\link{meanAbsCor}}.
}

