% Generated by roxygen2 (4.0.2): do not edit by hand
\name{shuffleSubjects}
\alias{shuffleSubjects}
\title{Shuffle subject identifiers}
\usage{
shuffleSubjects(x)
}
\arguments{
\item{x}{A data frame with column \code{SID} containing subject identifiers.}
}
\value{
Same data frame with shuffled IDs.
}
\description{
Shuffle subject identifiers
}

