% Generated by roxygen2 (4.0.2): do not edit by hand
\name{simulateGeneData}
\alias{simulateGeneData}
\title{Generate simulated gene expresion data}
\usage{
simulateGeneData(n, gmeans, gvcov, indep = FALSE)
}
\arguments{
\item{n}{Number of rows (subjects) of data to generate}

\item{gmeans}{Mean expression for each gene}

\item{gvcov}{}

\item{indep}{whether to generate dependent (default=\code{FALSE}) or independent (\code{TRUE}) data.}
}
\value{
An n x 53 matrix, with subjects forming rows and genes forming columns.
}
\description{
Simply wraps a call to \code{\link{mvrnorm}} in package \code{MASS}.  If \code{indep} is \code{TRUE}, zeroes out all of the covariances in the matrix.
}
\seealso{
\code{\link{generateVCMatrix}}
}

