% Generated by roxygen2 (4.0.2): do not edit by hand
\name{simulatePsychometricData}
\alias{simulatePsychometricData}
\title{Randomly generate psychometric/demographic data}
\usage{
simulatePsychometricData(n, pmeans, pvcov)
}
\arguments{
\item{n}{number of subjects (rows)}

\item{pmeans}{named vector of predictor means}

\item{pvcov}{variance-covariance matrix for predictors}
}
\value{
a dataframe of new predictors
}
\description{
Basically wraps a call to \code{\link{mvrnorm}}, with
some additional processing for categorical variables
}

