% Generated by roxygen2 (4.0.2): do not edit by hand
\name{oneSampTest}
\alias{oneSampTest}
\title{Perform a one sample t-test on the beta values}
\usage{
oneSampTest(x)
}
\arguments{
\item{x}{data frame, with beta values for one of the two psychometric factors in variable named \code{Y}}
}
\value{
data frame with results of one-sample test.
}
\description{
Perform a one sample t-test on the beta values
}

